+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # develop
 
+# 1.0.1 - 2012-04-25
+  * Bugfix: Actions which do not receive a response within 10s will allow further actions to be executed. Synchronous originate has a 60s timeout.
+
 # 1.0.0 - 2012-03-09
   * Bugfix: Remove rcov
   * Bump to 1.0.0 since we're in active use